崔思思的最新

+++ routes/web.php
@@ -1,4 +1,4 @@
-<?php
+ <?php
 
 /*
 |--------------------------------------------------------------------------
@@ -13,20 +13,20 @@
 
 
 
-//前台订单
+//前台订单  
+
 //下订单
 Route::get('home/ordersub/{gid}/{num}', 'Home\OrdersubController@index');	
 Route::get('home/ordersub', 'Home\OrdersubController@end');	
 	
 //生成订单
-Route::get('home/ordersub/create', 'Home\OrdersubController@create');			
-
+Route::post('home/ordersub', 'Home\OrdersubController@create');			
 //订单详情
 //Route::get('home/ordersinfo/{code}', 'Home\OrdersinfoController@index');			
 //取消订单
 //Route::get('home/orderre', 'Home\UserorderController@create'); 		
 //确认收货
-//Route::get('home/ordersu', 'Home\UserorderController@store'); 	
+//Route::get('home/ordersu', 'Home\UserorderController@store'); 		
 
 
 //前台
@@ -53,8 +53,6 @@ Route::get('home/goods/{id}','Home\GoodsController@show');
 // 前台路由组中间件
 Route::group(['prefix'=>'index','namespace'=>'Index','middleware'=>'index.login'],function(){
 
-	// 前台退出
-	Route::get('/index/user/loginout','Index\LoginController@loginout');
 
 
 	// 前台退出
@@ -87,18 +85,23 @@ Route::group(['prefix'=>'index','namespace'=>'Index','middleware'=>'index.login'
 	Route::post('/address/update','Address\AddressController@update');
 	Route::get('/address/delete/{id}','Address\AddressController@delete');
 
-	Route::get('/user/addshou/{id}','Shou\ShouController@add');
 
-	
+	//用户个人订单
+	Route::get('/user/order/{id}','Order\OrderController@order');
+
+
 });
 
 
 	//加入购物车
 	Route::post('/home/addcart','Home\ShopcarController@addCart');
+Route::group(['middleware'=>'index.login'],function(){
+
 	//购物车列表
 	Route::get('/home/shopcar','Home\ShopcarController@cart');
 	//删除商品
-	Route::post('/home/shopcar/delcart','Home\ShopcarController@delcart');
+	Route::post('//home/shopcar/delcart','Home\ShopcarController@delcart');
+});
 
 
 // 后台登录
@@ -110,19 +113,11 @@ Route::post('admin/dologin','Admin\LoginController@dologin');
 Route::group(['prefix'=>'admin','namespace'=>'Admin','middleware'=>'admin.login'],function(){
 
 	//后台首页
-Route::get('index','indexController@index');
-Route::group(['middleware'=>'hasRole'],function(){
+	Route::get('index','indexController@index');
 
 	// 后台退出
 	Route::get('loginout','LoginController@loginout');
 
-
-
-	//后台商品图片
-	Route::get('/goods/picture/{id}','GoodsController@picture');
-	Route::post('/goods/addpic','GoodsController@addpic');
-});
-         
 	//用户管理
 	Route::get('user/test','UsersController@test');
 	Route::get('user/address','UsersController@address');
@@ -130,9 +125,12 @@ Route::group(['middleware'=>'hasRole'],function(){
 	Route::resource('users','UsersController');
 	//    用户授权页面
 	Route::get('user/auth/{id}','UsersController@auth');
-	// 添加用户授权逻辑
+	//    添加用户授权逻辑
 	Route::post('user/doauth','UsersController@doauth');
-            
+
+
+
+
 	//角色相关的路由
 	Route::resource('role','RoleController');
 
@@ -142,6 +140,7 @@ Route::group(['middleware'=>'hasRole'],function(){
 	Route::get('role/auth/{id}','RoleController@auth');
 	//    添加用户授权逻辑
 	Route::post('role/doauth','RoleController@doauth');
+
 
 	//后台订单管理
 	Route::get('orders/up/{id}','OrdersController@up');
@@ -175,20 +174,20 @@ Route::group(['middleware'=>'hasRole'],function(){
 	Route::resource('/friendlink','FriendlinkController');
 
 	//分类
-	Route::get('cate/add','cateController@add');
-	Route::get('cate/edit/{id}','cateController@edit');
-	Route::delete('cate/del/{id}','cateController@del');
-	Route::put('cate/update/{id}','cateController@update');
-	Route::post('cate/changeOrder/{id}','cateController@changeOrder');
-	Route::post('config/changeOrder/{id}','ConfigController@changeOrder');
-	Route::post('cate','cateController@store');
+	Route::get('admin/cate/add','Admin\cateController@add');
+	Route::get('admin/cate/edit/{id}','Admin\cateController@edit');
+	Route::delete('admin/cate/del/{id}','Admin\cateController@del');
+	Route::put('admin/cate/update/{id}','Admin\cateController@update');
+	Route::post('admin/cate/changeOrder/{id}','Admin\cateController@changeOrder');
+	Route::post('admin/config/changeOrder/{id}','Admin\ConfigController@changeOrder');
+	Route::post('admin/cate','Admin\cateController@store');
 
 
-	Route::get('cate/list','cateController@index');
+	Route::get('admin/cate/list','Admin\cateController@index');
 
 	//网站配置模块
-	Route::resource('/config','ConfigController');
-	Route::post('config/changeContent','ConfigController@changeContent');
+	Route::resource('admin/config','Admin\ConfigController');
+	Route::post('admin/config/changeContent','Admin\ConfigController@changeContent');
 
 
 	//推荐位
